move common options definition to archiver.common
